optimizing the code - auth/user routes

+++ controllers/boardController.js
@@ -1,11 +1,19 @@
 const Board = require("../models/Board");
+const User = require("../models/User");
 
 exports.createBoard = async (req, res) => {
   try {
+    const { userId } = req.user;
+    const user = await User.findById(userId);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
     const board = await Board.create({
       title: req.body.title,
       userId: req.user.userId,
     });
+    user.boards.push(board);
+    await user.save();
     res.status(201).json(board);
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -14,8 +22,8 @@ exports.createBoard = async (req, res) => {
 
 exports.getSingleBoard = async (req, res) => {
   try {
-    const { boardId } = req.params;
-    const board = await Board.findById(boardId).populate("lists");
+    const { id } = req.params;
+    const board = await Board.findById(id).populate("lists");
     if (!board) {
       return res.status(404).json({ message: "Board not found" });
     }
@@ -27,9 +35,12 @@ exports.getSingleBoard = async (req, res) => {
 
 exports.getBoards = async (req, res) => {
   try {
-    const boards = await Board.find({ userId: req.user.userId }).populate(
-      "lists"
-    );
+    const { userId } = req.user;
+    const user = await User.findById(userId);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+    const boards = await Board.find({ userId }).populate("lists");
     res.json(boards);
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -38,10 +49,15 @@ exports.getBoards = async (req, res) => {
 
 exports.updateBoard = async (req, res) => {
   try {
-    const board = await Board.findByIdAndUpdate(req.params.id, req.body, {
+    const { id } = req.params;
+    const board = await Board.findById(id);
+    if (!board) {
+      return res.status(404).json({ message: "Board not found" });
+    }
+    updatedBoard = await Board.findByIdAndUpdate(req.params.id, req.body, {
       new: true,
     });
-    res.json(board);
+    res.json(updatedBoard);
   } catch (error) {
     res.status(400).json({ error: error.message });
   }
@@ -49,6 +65,11 @@ exports.updateBoard = async (req, res) => {
 
 exports.deleteBoard = async (req, res) => {
   try {
+    const { id } = req.params;
+    const board = await Board.findById(id);
+    if (!board) {
+      return res.status(404).json({ message: "Board not found" });
+    }
     await Board.findByIdAndDelete(req.params.id);
     res.json({ message: "Board deleted" });
   } catch (error) {
